ANDROID: cpufreq: concurrent_active_time by pid
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each pid spent running
with 1 to num_possible_cpus online and not idle.

We can use this information to blame pids and
uids for the energy they are using.

For instance, if 4 cores are online and active
for 50ms and the pid is running on one of the
cores, the pid should be blamed for 1/4 of the
base energy used by the cpu when 1 or more cores is
active. Using other proc files, the pid can
also be blamed for the cost of keeping a cluster
on and the cost of the individual cores it
is using.

b/68399339

Test: cat /proc/<PID>/concurrent_active_time

Change-Id: I5262988425aa3d8f54565f6be10a0af59eee5d41
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>